Add weblate schema localization pull support
