+++ specifyweb/frontend/js_src/lib/localization/schema-localization/toPoFile.ts
@@ -1,10 +1,11 @@
 import { f } from '../../utils/functools';
 import type { RA } from '../../utils/types';
-import { camelToHuman, capitalize } from '../../utils/utils';
+import { capitalize } from '../../utils/utils';
 import type { LocalizationEntry } from '../utils';
 import type { Language } from '../utils/config';
 import type { DictionaryUsages } from '../utils/scanUsages';
 import { syncStrings } from '../utils/sync';
+import { schemaLocalizationFile } from './gatherLocalization';
 import type { SchemaStrings } from './parser';
 import type { SchemaLocation } from './traversal';
 
@@ -15,10 +16,11 @@ import type { SchemaLocation } from './traversal';
 export const schemaToPoFile = async (
   structure: SchemaStrings,
   languages: RA<string>,
+  weblateDirectory: string,
   sourceFilePath: string
 ): Promise<DictionaryUsages[string]> => {
   const dictionaryUsages = {
-    categoryName: pathToName(sourceFilePath),
+    categoryName: schemaPathToComponentName(sourceFilePath),
     strings: Object.fromEntries(
       Object.entries(structure).map(
         ([key, { strings, locations }]) =>
@@ -44,7 +46,7 @@ export const schemaToPoFile = async (
       usage: ({ filePath }) => filePath,
       reference: f.undefined,
     },
-    './'
+    weblateDirectory
   );
 
   return dictionaryUsages;
@@ -61,10 +63,29 @@ const mainSchemaName = 'main';
  * differentiate them
  */
 export const schemaLocalizationName = 'schema-localization-';
-const pathToName = (sourceFilePath: string): string =>
-  `${schemaLocalizationName}${camelToHuman(
-    sourceFilePath.includes('/') ? sourceFilePath.split('/')[0] : mainSchemaName
-  )}`;
+const pathPathJoin = '-';
+
+function schemaPathToComponentName(sourceFilePath: string): string {
+  if (sourceFilePath.includes(pathPathJoin))
+    throw new Error(
+      'Dash is not supported in path at the moment. ' +
+        'Modify componentNameToSchemaPath if need to support them'
+    );
+
+  const parts = sourceFilePath.split('/').slice(0, -1);
+  return `${schemaLocalizationName}${
+    parts.length === 0 ? mainSchemaName : parts.join(pathPathJoin)
+  }`;
+}
+
+export function componentNameToSchemaPath(componentName: string): string {
+  if (!componentName.startsWith(schemaLocalizationName))
+    throw new Error('Invalid schema localization component name');
+  const trimmedName = componentName.slice(schemaLocalizationName.length);
+  const parts =
+    trimmedName === mainSchemaName ? [] : trimmedName.split(pathPathJoin);
+  return [...parts, schemaLocalizationFile].join('/');
+}
 
 /**
  * Convert a place where the string is used to a user-friendly string
@@ -75,3 +96,9 @@ const locationToString = (location: SchemaLocation): string =>
       ? ''
       : `- ${capitalize(location.fieldName)}`
   } [${capitalize(location.type)}]`;
+
+export const exportsForTests = {
+  schemaPathToComponentName,
+  mainSchemaName,
+  locationToString,
+};
