+++ specifyweb/frontend/js_src/lib/utils/__tests__/utils.test.ts
@@ -141,7 +141,6 @@ test('multiSortFunction', () => {
     ].sort(
       multiSortFunction(
         ({ type }) => type,
-        false,
         ({ priority }) => priority,
         true
       )
